ideation tools new spec
