+++ tools/IdeationTools/__init__.py
@@ -1,0 +1,3 @@
+from .tool import Tool
+
+__all__ = ["Tool"]
